QoL - add token size and image scale to token folder override options

+++ SidebarPanel.js
@@ -545,20 +545,20 @@ function build_rangeInput_input(settingOption, currentValue, changeHandler){
     
     range.on('input change', function(){
    
-      const rangeValue = parseInt(range.val());
+      const rangeValue = parseFloat(range.val());
       input.val(rangeValue);
       changeHandler(settingOption.name, rangeValue);
     });
 
 
     range.on('mouseup', function(){
-        const rangeValue = parseInt(range.val());
+      const rangeValue = parseFloat(range.val());
         changeHandler(settingOption.name, rangeValue);
     });
 
     input.on('input change', function(){
       $("#darkness_layer").toggleClass("smooth-transition", true);
-      const inputValue = parseInt(input.val());
+      const inputValue = parseFloat(input.val());
       range.val(inputValue);
       changeHandler(settingOption.name, inputValue);
     });
